Create SECURITY.md Habib Hossain
By submitting the personal information above, I acknowledge that my personal information is being collected under the privacy policy of the business from which I am seeking a discount, and I understand that my personal information will be shared with SheerID as a processor/third-party service provider in order for SheerID to confirm my eligibility for a special offer. More about SheerID.Your information will be transferred to the United States and will be treated according to SheerID’s privacy policy.